Add files via upload
A program which uses parallel arrays to store information found on each line of an input file. Each line contains a bowling team member, the name of their team, and the team members score. The program sums the scores of the teams that are listed and after the maximum score is determined, the program will display the team who won and then displays the winning team members names and scores