Add files via upload
Java program using a GUI that allows users to enter an assessed value of their home, the school tax rate, and the county tax rate. After the calculate button is clicked, the assessed home value and the two tax rates are then used to calculate the school and county taxes as well as the total taxes. When the exit JButton is clicked the program shuts down.