Add files via upload
GUI created using Java that allows a user to enter an assessed home value, the school tax rate, and the county tax rate. After the calculate button is clicked by the user, the assessed home value and the two tax rates are used to calculate the school, and county taxes as well as the toatl taxes. When the exit JButton is clicked, the program will exit.